renamed GrimBoatLongJump.kt to VehicleLongJump.kt and made it actually work
