+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/movement/longjump/ModuleLongJump.kt
@@ -25,8 +25,8 @@ import net.ccbluex.liquidbounce.event.events.PlayerJumpEvent
 import net.ccbluex.liquidbounce.event.handler
 import net.ccbluex.liquidbounce.features.module.Category
 import net.ccbluex.liquidbounce.features.module.Module
+import net.ccbluex.liquidbounce.features.module.modules.movement.longjump.modes.VehicleLongJump
 import net.ccbluex.liquidbounce.features.module.modules.movement.longjump.modes.VulcanLongJump
-import net.ccbluex.liquidbounce.features.module.modules.movement.longjump.modes.grim.GrimBoatLongJump
 import net.ccbluex.liquidbounce.features.module.modules.movement.longjump.modes.nocheatplus.NoCheatPlusBoost
 import net.ccbluex.liquidbounce.features.module.modules.movement.longjump.modes.nocheatplus.NoCheatPlusBow
 import net.ccbluex.liquidbounce.utils.entity.moving
@@ -43,7 +43,7 @@ object ModuleLongJump : Module("LongJump", Category.MOVEMENT) {
             NoCheatPlusBoost,
             NoCheatPlusBow,
             VulcanLongJump,
-            GrimBoatLongJump
+            VehicleLongJump
         )
     )
     private val autoJump by boolean("AutoJump", false)
